pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kalshi-autotrader"
version = "0.1.0"
description = "Automated trading system for Kalshi prediction markets, focused on AI leaderboard contracts"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "kalshi-python>=2.1.4",
    "websockets>=13.0",
    "sqlalchemy>=2.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "apscheduler>=3.10",
    "structlog>=24.0",
    "httpx>=0.27",
    "discord.py>=2.3",
    "beautifulsoup4>=4.12",
    "pyyaml>=6.0",
    "rich>=13.0",
    "thefuzz[speedup]>=0.22",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-PyYAML",
    "types-beautifulsoup4",
]

[project.scripts]
autotrader = "autotrader.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/autotrader"]

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "kalshi_python.*",
    "thefuzz.*",
    "apscheduler.*",
    "discord.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
